requirements.txt
pytest==7.4.0
pytest-playwright==0.4.0
playwright==1.39.0
pytest-html==4.0.2
pytest-xdist==3.3.1
allure-pytest==2.13.2
pytest-rerunfailures==12.0
python-dotenv==1.0.0
pyyaml==6.0.1
faker==24.4.0 